% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elastic_donut_sp}
\alias{elastic_donut_sp}
\title{Estimate torus model using elasticnet}
\usage{
elastic_donut_sp(locus, z, X, EM_thresh = 0.05, alpha = 0,
  lambda = as.numeric(c(0)), prior_init = 0.001)
}
\arguments{
\item{X}{}
}
\description{
Estimate torus model using elasticnet
}
