% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{forward_op_torus_cmd}
\alias{forward_op_torus_cmd}
\title{Title}
\usage{
forward_op_torus_cmd(gf, anno_df, f_feat_df, term_df, i, p,
  prior = NA_integer_, verbose = F)
}
\arguments{
\item{gf}{}

\item{anno_df}{}

\item{f_feat_df}{}

\item{term_df}{}

\item{i}{}

\item{p}{}

\item{prior}{}

\item{verbose}{}
}
\description{
Title
}
