% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{forward_select_fun}
\alias{forward_select_fun}
\title{Perform a 'forward selection'}
\usage{
forward_select_fun(f, params, combo_fun, extract_terms, steps = 1L,
  ret_all = FALSE)
}
\arguments{
\item{f}{A function that fits the underlying model, taking \code{X} (and anything passed to \code{...}) as argument.
\code{f} should return a list with a \code{lik} element, which will be used for the forward selection.}

\item{params}{a list over which to perform the forward selection.}

\item{combo_fun}{a function for generating input to \code{f} by combining elements of \code{params}.  \code{combo_fun} should
accept subsets of \code{params} that are of length 0(for the NULL model) or greater (for subsequent cases). At each step of
the forward selection, each remaining element (not in the current model) of \code{params} will be added to the model
using \code{combo_fun},the results of which will be passed to \code{f}}

\item{extract_terms}{a function for extracting the relevant terms from the model, so the params\link{extract_terms}}
}
\description{
Perform a 'forward selection'
}
