% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{fs_torus}
\alias{fs_torus}
\title{Title}
\usage{
fs_torus(gwas_df, full_anno_df, steps = 1L, p_cutoff = 1,
  torus_p = character(0))
}
\arguments{
\item{gwas_df}{}

\item{full_anno_df}{}

\item{steps}{}

\item{p_cutoff}{}

\item{torus_p}{}
}
\description{
Title
}
