% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{spread_sparse}
\alias{spread_sparse}
\title{Make sparse annotation matrix from annotation dataframe}
\usage{
spread_sparse(anno_df, SNP = "SNP", feature = "feature",
  row_total = length(unique(anno_df$SNP)))
}
\arguments{
\item{anno_df}{dataframe with columns}

\item{row_total}{total number of SNPs/loci}
}
\value{
dgCMatrix with features as columns and loci as rows
}
\description{
Make sparse annotation matrix from annotation dataframe
}
