% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{write_anno}
\alias{write_anno}
\title{Title}
\usage{
write_anno(anno_df = tibble(SNP = integer(), feature = character()),
  p = max(max(anno_df$SNP), 1L), af = tempfile(fileext = ".txt.gz"))
}
\arguments{
\item{anno_df}{}

\item{p}{}

\item{af}{}
}
\description{
Title
}
